refactor : fold zone-navigation older/post-EOT tests into one file
Merges test_zone_agent_navigation_older_enemies_e2e.py and
test_zone_agent_navigation_post_eot_e2e.py into the base
test_zone_agent_navigation_e2e.py with two class-scoped fixtures that
chain on top of the module fixture's fresh TestConfig load.

TestZoneBoxAfterFirstEot recruits a Charlie/Echo double via the
perfect-worker form, runs one EOT, then covers the doubles fold (Echo
view), the recent enemies list (Alpha view), and the absence of a Plus
anciens fold after a single EOT.

TestZoneBoxAfterAgingCke moves Searcher_1 out and runs three more EOTs
so the initial CKE entries age past attackTimeWindow=1 and the Plus
anciens fold appears.

Three files into one; one TestConfig load instead of three saves
roughly a minute of suite wall-clock.
